pyproject.toml
[tool.poetry]
name = "bulletin-board-system"
version = "0.1.0"
description = ""
authors = ["Sam Gabo <sam.gabo@bld.ai>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"
django-extensions = "^3.2.1"
ipython = "^8.5.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
mypy = "^0.982"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
